reorg and shipping
