+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Checkout/Checkout+Internal.swift
@@ -8,8 +8,26 @@
 import Foundation
 @_spi(STP) import StripeCore
 @_spi(STP) import StripePayments
+import StripeApplePay
 
-extension Checkout: ExpressCheckoutElementDelegate {}
+extension Checkout: ExpressCheckoutElementDelegate {
+    /// Called by ExpressCheckoutElement when the user taps the Apple Pay button.
+    func confirmApplePay() {
+        let confirmHandler = configuration.expressCheckoutElement.confirmHandler
+        guard let context = STPApplePayContext.create(checkout: self, confirmHandler: { result in
+            confirmHandler?(result)
+        }) else {
+            confirmHandler?(.failed(CheckoutError.applePayNotSupportedOrMisconfigured))
+            return
+        }
+        Task { @MainActor [weak self] in
+            guard let self else { return }
+            await enqueueSessionUpdate {
+                context.presentApplePay()
+            }
+        }
+    }
+}
 extension Checkout: CurrencySelectorElementDelegate {}
 
 extension Checkout {
